requirements.txt
aiohttp>=3.8.3
beautifulsoup4>=4.11.1
discord.py>=2.0.1
howlongtobeatpy>=1.0.2
Pillow>=9.3.0
python-dotenv>=0.21.0
python_dateutil>=2.8.2
pytz>=2022.6
wavelink>=1.3.3